Finalize private identity boundaries
