+++ tach.toml
@@ -97,10 +97,18 @@ depends_on = ["mindroom.durable_write"]
 
 [[modules]]
 path = "mindroom.private_instance_identity"
+depends_on = ["mindroom.private_instance_identity_store"]
+
+[[modules]]
+path = "mindroom.private_instance_identity_store"
 depends_on = [
     "mindroom.durable_write",
     "mindroom.file_locks",
     "mindroom.tool_system.worker_routing",
+]
+visibility = [
+    "mindroom.private_instance_identity",
+    "mindroom.runtime_resolution",
 ]
 
 [[modules]]
@@ -801,7 +809,7 @@ path = "mindroom.runtime_resolution"
 depends_on = [
     "mindroom.agent_policy",
     "mindroom.constants",
-    "mindroom.private_instance_identity",
+    "mindroom.private_instance_identity_store",
     "mindroom.tool_system.worker_routing",
     "mindroom.workspaces",
 ]
@@ -4171,7 +4179,6 @@ from = ["mindroom.runtime_resolution"]
 expose = [
     "PrivateInstanceIdentity",
     "PrivateInstanceIdentityError",
-    "ensure_private_instance_identity",
     "load_private_instance_identity",
 ]
 from = ["mindroom.private_instance_identity"]
